Implemented the state handling for the view pager container fragment

+++ app/src/main/java/com/example/pecodetask/features/pageContainer/model/PagerItem.kt
@@ -1,5 +1,5 @@
 package com.example.pecodetask.features.pageContainer.model
 
 data class PagerItem(
-    val pageNumber: Int
+    val pageNumber: Long
 )
